[💎 : refactor] 컴포넌트 레포 정리 및 import 구문 통일
