+++ src/App.tsx
@@ -3,8 +3,8 @@ import { RouterProvider, createBrowserRouter } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
 import { ThemeProvider } from 'styled-components';
 import { auth } from '@/firebase.ts';
-import LoadingSpinner from '@compo/loading-spinner.tsx';
-import ModeToggle from '@compo/mode-toggle.tsx';
+import LoadingSpinner from '@compo/common/loading-spinner.tsx';
+import ModeToggle from '@compo/common/mode-toggle.tsx';
 import isDarkAtom from '@atom/is-dark.tsx';
 import { lightTheme, darkTheme } from '@style/theme.ts';
 import GlobalStyles from '@style/global.ts';
