+++ src/components/common/window-top.tsx
