+++ src/components/left-side-menu.tsx
@@ -1,4 +1,4 @@
-import WindowTop from '@compo/window-top.tsx';
+import WindowTop from '@compo/common/window-top.tsx';
 import MiniProfile from '@compo/left-side-menu/mini-profile.tsx';
 import Navigation from '@compo/left-side-menu/navigation.tsx';
 import * as S from '@style/window.ts';
