+++ src/components/profile/user-profile.tsx
@@ -5,10 +5,10 @@ import { deleteUser } from 'firebase/auth';
 import { deleteDoc, doc } from 'firebase/firestore';
 import { deleteObject, ref } from 'firebase/storage';
 import { auth, db, storage } from '@/firebase.ts';
+import EditProfileForm from '@compo/profile/edit-profile-form.tsx';
 import currentUserAtom from '@atom/current-user.tsx';
 import IUser from '@type/IUser.ts';
 import useErrorMessage from '@hook/useErrorMessage.tsx';
-import EditProfileForm from '@compo/profile/edit-profile-form.tsx';
 import * as S from '@style/profile.ts';
 import * as P from '@style/popup.ts';
 import ErrorAlarm from '@style/error-alarm.ts';
