+++ src/routes/home.tsx
@@ -1,4 +1,4 @@
-import WindowTop from '@compo/window-top.tsx';
+import WindowTop from '@compo/common/window-top.tsx';
 import PostTweetForm from '@compo/home/post-tweet-form.tsx';
 import Timeline from '@compo/home/timeline.tsx';
 import * as W from '@style/window.ts';
