+++ src/routes/search-result.tsx
@@ -1,5 +1,5 @@
+import WindowTop from '@compo/common/window-top.tsx';
 import SearchTimeline from '@compo/search-result/search-timeline.tsx';
-import WindowTop from '@compo/window-top.tsx';
 import * as W from '@style/window.ts';
 
 export default function SearchResult() {
